requirements.txt
certifi==2020.12.5
chardet==3.0.4
filestack-python==3.2.1
fpdf==1.7.2
idna==2.10
requests==2.23.0
trafaret==2.0.2
urllib3==1.25.11
